as4610: don't fail if boot_diag does not exist in environment
The boot_diag variable may not exist. In this case, we do not need to
fixup anything, but the install will be aborted because fw_printenv will
return failure.

Fix this by always succeeding. If the u-boot environment is broken, the
previous fw_printenv command will already have caught that.

Reported-by: Jan Klare <jan.klare@bisdn.de>
Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>