installer: add backup library functions
Add code for new style of backup/restore, that makes use of OPKG's
conffiles and custom lists of files to (not) back up.

The supported syntax for files is #* for comments, /* for files to
backup and -/* to negate previous backup directives.

No support for backing up individual symlinks due to limitations in
ONIE's busybox environment.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>