installer: setup ONLP at install time
Populate /etc/onl/platform and the libonl-platform.so.1 link at install
time based on ONIE info. As a safeguard, only update the
libonlp-platform link if the appropriate platform library exists.

For now do not emit an error message if the library does not exist, as
accton-as4610 images only contain the as4610-30 library. It supports the
-54 as well, but technically there is no as4610-54 library.

Try both libonlp-<arch>-<platform>-<revision>.so as well as without
revision, loosely matching what ONL does [1].

This will allow supporting multiple machines with one image in the
future.

[1] https://github.com/opencomputeproject/OpenNetworkLinux/blob/master/packages/base/all/vendor-config-onl/src/python/onl/platform/baseconfig.py#L50

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>