backup: make sure default sudoers backup files are ignored
Sudo has special rules for the filename to be ignored, only files with
names ending in ~ (like created by vim) or files with a . in the name
are ignored.

So add a check for the path of the file and make sure we use .default
instead of -default for files in /etc/sudoers.d/.

Fixes changes in /etc/sudoers.d/basebox not sticking through
sysupgrades, due to the basebox-default being followed by sudo.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>