backup: make sure /etc/gshadow is taken over as well
Due to an oversight releases before 4.7.0 do not contain /etc/gshadow in
their default system backup list. This can cause issues for groups that
are present in one file but not the other, causing e.g. systemd to fail
to upgrade older groups:

             Starting Create System Users...
    [FAILED] Failed to start Create System Users.

Fix this by ensuring we always take over /etc/gshadow when backing up
/etc/groups.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>